refactor(broker): step 4/6 — extract the three comparator modules
Step 4 of the directory split. Pulls every comparator out of
replay-check.ts and into three cohesive modules:

  cost-ledger/replay-check/
  ├── crossing-comparators.ts (~318 LOC)
  │     - compareCrossings  (projection vs log)
  │     - detectDuplicateLoggedCrossings
  │     - detectCausalOrderViolations
  │     - validateBudgetSetReferences
  │     - detectDelayedEmissions
  │     - validateLoggedCrossingReferences
  │     - compareExpectedAndLoggedCrossings  (oracle vs log)
  ├── aggregate-comparators.ts (~86 LOC)
  │     - compareAgentDays
  │     - compareTasks
  └── budget-comparator.ts (~100 LOC)
        - compareBudgets

The grouping mirrors the discrepancy-variant taxonomy already split
across the union in discrepancy.ts: each module owns the comparators
that emit a related cluster of `threshold_crossing_*`, `agent_day_*`/
`task_*`, or `budget_*` variants. replay-check.ts is now the
orchestrator file only.

replay-check.ts drops from 919 LOC to 446 LOC.

All 249 broker tests pass; typecheck clean.
